QVAC-21225 test: extract + harden flaky HTTP test servers
- Move FlakyFileServer + ShardSeverProxy out of the node executor into
  tests/shared/flaky-http-servers.ts so they aren't inlined in the executor
- Minimize FlakyFileServer: send one partial then drop (auto) or hold until
  sever() (manual), dropping the timed trickle pump — simplest tool for the test
- ShardSeverProxy: resolve the upstream from a fixed shard->URL table keyed by
  request basename, so no request-derived string reaches fetch(), and return a
  generic text/plain error body instead of echoing the exception — clears the
  CodeQL SSRF + exception-exposure findings
- Desktop validated: registry-suspend, http-netdrop, http-suspend and (gated)
  http-sharded all green after the change
